@JoanFM 2023-07-24
Can you also add how much more memory is used? I am not sure this is desired.