@jupyterjazz 2023-07-24
@JoanFM 

Not much change in terms of memory too.

Here's the experiment code
```python
from docarray.index import InMemoryExactNNIndex
from docarray import BaseDoc
from docarray.typing import NdArray
import numpy as np
import time
import tracemalloc

tracemalloc.start()

class MyDoc(BaseDoc):
    text: str
    embedding: NdArray[128]


data = [MyDoc(text=f'text {i}', embedding=np.random.rand(128)) for i in range(200000)]
doc_index = InMemoryExactNNIndex[MyDoc]()

doc_index.index(data)

start_find = time.time()
docs, scores = doc_index.find(data[0], search_field='embedding')
end_find = time.time()

if data[10] in data:
    print('wohoo')

ids_to_get = [data[200].id, data[250].id, data[350].id]

ids_to_del = [data[200].id, data[250].id, data[350].id]
del doc_index[ids_to_del]


print(tracemalloc.get_traced_memory())

tracemalloc.stop()
```
Before change prints (768773718, 1032141402) showing current memory usage and max memory usage
After change - (784741328, 1048109012)
So there's ~2% increase